feat: Update footer links and GitHub profile
This commit updates the footer links in the Footer component to point to Bichitra's blog instead of Sahand's blog. It also updates the GitHub profile link to point to Bichitra's GitHub profile instead of Sahand's GitHub profile. This change ensures that the correct links are displayed and improves the accuracy of the information provided to users.