chore: Update Header component layout and link styles
This commit updates the Header component layout by adding a new div container to group the search form and search button. It also improves the link styles by adding the "text-inherit" and "no-underline" classes to the Link components. This change enhances the user experience by providing a more organized layout and improving the visual appearance of the links in the header.