chore: Update npm dependencies
This commit updates the npm dependencies in the package.json file. Specifically, it updates the versions of the following dependencies:
- firebase: from 10.11.1 to 10.12.0
- react-router-dom: from 6.23.0 to 6.23.1
- @types/react: from 18.3.1 to 18.3.2

By updating these dependencies, we ensure that our project is using the latest stable versions, which may include bug fixes, performance improvements, and new features. This update helps maintain the stability and compatibility of our application.